split project into workspaces
