+++ .github/workflows/build.yml
@@ -1,0 +1,20 @@
+on: [push]
+
+name: CI
+
+jobs:
+  build:
+    name: Rust project
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout 🛎
+        uses: actions/checkout@v2
+      - name: Setup Rust 🦀
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+      - name: Build projects 🚧
+        uses: actions-rs/cargo@v1
+        with:
+          command: build
+          args: --release --all-features
