+++ Cargo.toml
@@ -1,12 +1,7 @@
-[package]
-name = "lgtm"
-version = "0.1.0"
-authors = ["Will Toher <will.toher.0@gmail.com>"]
-edition = "2018"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-actix-rt = "1.0"
-actix-web = "2.0"
-rand = "0.7.3"
+members = [
+    "api",
+    "lgtm",
+    "cli",
+]
