+++ api/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "api"
+version = "0.1.0"
+authors = ["Will Toher <will.toher.0@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+actix-rt = "1.0"
+actix-web = "2.0"
+lgtm = { path = "../lgtm" }
