+++ cli/src/main.rs
@@ -1,0 +1,5 @@
+use lgtm;
+
+fn main() {
+    println!("{}", lgtm::lgtm());
+}
