+++ lgtm/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "lgtm"
+version = "0.1.0"
+authors = ["Will Toher <will.toher.0@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rand = "0.7.3"
