+++ lgtm/src/dicts/g.txt
