+++ lgtm/src/dicts/l.txt
