+++ lgtm/src/dicts/m.txt
