+++ lgtm/src/dicts/t.txt
